refactor: update error handling for cd; fix: add error message const for ls
